Simple dictionaries phase 1

+++ src/Models.fs
@@ -99,11 +99,11 @@ type AccountUpd() =
 
 type Account() =
         [<DefaultValue>]val mutable id: int
-        [<DefaultValue>]val mutable fname : int64
-        [<DefaultValue>]val mutable sname : int64
+        [<DefaultValue>]val mutable fname : byte
+        [<DefaultValue>]val mutable sname : int16
         [<DefaultValue>]val mutable email: string
         [<DefaultValue>]val mutable emailDomain: string
-        [<DefaultValue>]val mutable interests: int64[]
+        [<DefaultValue>]val mutable interests: byte[]
         [<DefaultValue>]val mutable status: byte
         [<DefaultValue>]val mutable premiumStart: int
         [<DefaultValue>]val mutable premiumFinish: int
@@ -116,8 +116,8 @@ type Account() =
         [<DefaultValue>]val mutable likes: ResizeArray<int>
         [<DefaultValue>]val mutable joined: int
         [<DefaultValue>]val mutable joinedYear: int16
-        [<DefaultValue>]val mutable city: int64
-        [<DefaultValue>]val mutable country: int64
+        [<DefaultValue>]val mutable city: int16
+        [<DefaultValue>]val mutable country: byte
 
         member this.CreateCopy() =
             let copy = Account()
